Controllers: Back: add latest policies

+++ app/Http/Controllers/Back/AdminCosplayerController.php
@@ -21,9 +21,11 @@ class AdminCosplayerController extends Controller
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
+        $this->authorize('index', Cosplayer::class);
         $cosplayers = Cosplayer::latest()->get();
 
         return view('admin.cosplayers.index', [
@@ -35,9 +37,11 @@ class AdminCosplayerController extends Controller
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize('create', Cosplayer::class);
         return view('admin.cosplayers.create');
     }
 
@@ -46,12 +50,16 @@ class AdminCosplayerController extends Controller
      *
      * @param CosplayerRequest $request
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(CosplayerRequest $request)
     {
-        $cosplayer = Cosplayer::create([
-            'name' => $request->input('name'),
-        ]);
+        $cosplayer = new Cosplayer();
+        $cosplayer->name = $request->input('name');
+
+        $this->authorize('create', $cosplayer);
+
+        $cosplayer->save();
 
         return redirect(route('admin.cosplayers.show', ['cosplayer' => $cosplayer]));
     }
@@ -61,9 +69,11 @@ class AdminCosplayerController extends Controller
      *
      * @param  \App\Models\Cosplayer $cosplayer
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(Cosplayer $cosplayer)
     {
+        $this->authorize('view', $cosplayer);
         return view('admin.cosplayers.show', ['cosplayer' => $cosplayer]);
     }
 
@@ -72,9 +82,11 @@ class AdminCosplayerController extends Controller
      *
      * @param  \App\Models\Cosplayer $cosplayer
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Cosplayer $cosplayer)
     {
+        $this->authorize('edit', $cosplayer);
         return view('admin.cosplayers.edit', ['cosplayer' => $cosplayer]);
     }
 
@@ -84,11 +96,14 @@ class AdminCosplayerController extends Controller
      * @param CosplayerRequest $request
      * @param $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(CosplayerRequest $request, $id)
     {
         //TODO Update categories
         $cosplayer = Cosplayer::find($id);
+
+        $this->authorize('update', $cosplayer);
 
         $cosplayer->name = $request->input('name');
         $cosplayer->save();
@@ -105,6 +120,7 @@ class AdminCosplayerController extends Controller
      */
     public function destroy(Cosplayer $cosplayer)
     {
+        $this->authorize('destroy', $cosplayer);
         $cosplayer->delete();
         return Redirect::back()->withSuccess('Cosplayer successfully deleted');
     }
